moving toward a compiled library, including fftw
