+++ .gitignore
@@ -1,3 +1,4 @@
+fftw*
 settings.json
 .vscode*
 build/
