+++ Waves2AMR.h
@@ -1,0 +1,8 @@
+#ifndef WAVES2AMR_H
+#define WAVES2AMR_H
+// All headers from src
+#include "src/read_modes.h"
+#include "src/modes_hosgrid.h"
+#include "src/hosgrid_spatial_interp.h"
+
+#endif
